Add files via upload
This program took me absolutely forever to make but I am very certain that it finally does exactly what you asked for in the project instructions.  I tested it out based on the sample on Brightspace given and it worked perfectly.  I also added as many comments as I thought were appropriate for your understanding if you wanted to look deeper at the code.  This is my second commit for the redo grade submission.